SONARPLUGINS-1691 Analysis is blocking sporadically
Call pylint via CommandExecutor which works around the stupid Runtime.exec issues.
Also, remove the whole environment manipulation when calling pylint and according
configuration options as those things are better placed on the site
(in a shell script or similar). Revive a couple of pylint test and do some kaizen
along the way.