S935: Fix ruleSpecification value in metadata
There was an "S" in front of the rule number in the "ruleSpecification" value
for the python version of the rule. I.e. "RSPEC-S935" -> "RSPEC-935".

This caused problems with the deployment of the rules marketing website:
http://rules.sonarsource.com/

The corresponding change in the rspec repository have already been made,
see the following PR: https://github.com/SonarSource/rspec/pull/2895
The changes of this commit have been generated using the rule API.