SONARPY-1895: Use SonarLintCache provided by analyzer-commons
The SonarLintCache component which is used by DBD to transfer IR
in-memory from a custom sonar-python rule to the DBD plugin is now also
being utilized to transfer IR from a sonar-java rule to DBD. Since the
component is now utilized by more than one plugin, it was decided to move
it into analyzer-commons.

Hence, the duplicated component can and should be eliminated from the
sonar-python codebase.